requirements.txt
astunparse==1.6.3
backcall==0.2.0
Bottleneck==1.3.5
certifi==2022.12.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
dominate==2.8.0
fonttools==4.38.0
imageio==2.31.1
ipdb==0.13.13
ipython==7.34.0
jedi==0.18.2
jittor==1.3.4.14
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numexpr==2.8.4
numpy>=1.21.6
opencv-python==4.8.0.76
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboardX==2.6
tifffile==2021.11.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
